requirements.txt
aiokafka==0.7.2
aiosqlite==0.17.0
anyio==3.6.1
asgiref==3.5.2
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
ecdsa==0.17.0
fastapi==0.78.0
h11==0.13.0
idna==3.3
kafka-python==2.0.2
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
PyQt3D==5.15.5
PyQt5==5.15.7
PyQt5-sip==12.11.0
PyQtChart==5.15.6
PyQtDataVisualization==5.15.5
PyQtNetworkAuth==5.15.5
PyQtPurchasing==5.15.5
PyQtWebEngine==5.15.6
python-jose==3.3.0
python-multipart==0.0.5
qasync==0.23.0
requests==2.28.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
websockets==10.3
